1475: add line info in parsing error, rename folder
